django/requirements.txt
Django>=1.8
MySQL-python>=1.2
pillow>=2.9
baidubce==0.8.3
uwsgi>=2.0
